backend/requirements.txt
asgiref==3.11.0
boto3==1.42.16
botocore==1.42.16
Django==6.0
django-cors-headers==4.9.0
django-model-utils==5.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jmespath==1.0.1
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.2
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
bleach==6.2.0
django-csp==3.8
